goods/requirements.txt
Django
djangorestframework
psycopg2
drf_yasg